Remove auth on all endpoints.
The frontend provides full credentials for development right now, so the auth control in the API is meaningless. Additionally, the switch to Amplify environments creates problems when authenticating on more than one user pool. Eventually we will need to find some solution to secure the application, potentially by bypassing the python Warrant library (which interfaces with cognito)...